rename: 더보기 버튼의 이름 변경
