+++ app/src/main/java/woowacourse/shopping/feature/main/MainActivity.kt
@@ -11,7 +11,7 @@ import woowacourse.shopping.data.product.ProductDummyRepository
 import woowacourse.shopping.databinding.ActivityMainBinding
 import woowacourse.shopping.feature.cart.CartActivity
 import woowacourse.shopping.feature.detail.ProductDetailActivity
-import woowacourse.shopping.feature.main.adapter.MainViewModelFactory
+import woowacourse.shopping.feature.main.viewmodel.MainViewModelFactory
 import woowacourse.shopping.feature.main.adapter.ProductAdapter
 import woowacourse.shopping.feature.main.viewmodel.MainViewModel
