+++ app/src/main/res/layout/activity_main.xml
@@ -43,13 +43,13 @@
             tools:listitem="@layout/item_product" />
 
         <Button
-            android:id="@+id/btn_main_see_more"
+            android:id="@+id/btn_main_load_more"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:backgroundTint="@color/teal_500"
             android:onClick="@{() -> viewModel.loadPage()}"
             android:paddingVertical="12dp"
-            android:text="@string/see_more"
+            android:text="@string/load_more"
             android:textColor="@color/white"
             android:textSize="20sp"
             android:textStyle="bold"
@@ -57,7 +57,7 @@
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_goneMarginTop="100dp"
-            app:visibility="@{viewModel.showSeeMore}" />
+            app:visibility="@{viewModel.showLoadMore}" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
