--- app/src/main/res/layout/item_see_more.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <data>
-
-    </data>
-
-    <Button
-        android:id="@+id/btn_main_see_more"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:backgroundTint="@color/teal_500"
-        android:paddingVertical="12dp"
-        android:text="@string/see_more"
-        android:textColor="@color/white"
-        android:textSize="20sp"
-        android:textStyle="bold" />
-</layout>
-
-
